@@ -8,7 +8,7 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^8.0.2",
+        "php": "^8.1",
         "calebporzio/sushi": "^2.4",
         "doctrine/dbal": "^3.0",
         "filament/filament": "^2.16",
@@ -22,7 +22,9 @@
         "saade/filament-fullcalendar": "^1.6"
     },
     "require-dev": {
+        "phpmd/phpmd": "^2.13.0",
         "barryvdh/laravel-debugbar": "^3.7",
+        "barryvdh/laravel-ide-helper": "^2.13",
         "fakerphp/faker": "^1.9.1",
         "laravel/breeze": "^1.16",
         "laravel/pint": "^1.0",
