@@ -6,48 +6,57 @@ use App\Filament\Resources\KategoriResource\Pages;
 use App\Filament\Resources\KategoriResource\RelationManagers;
 use App\Models\Kategori;
 use Filament\Forms;
+use Filament\Forms\Components\TextInput;
 use Filament\Resources\Form;
 use Filament\Resources\Resource;
 use Filament\Resources\Table;
 use Filament\Tables;
+use Filament\Tables\Columns\TextColumn;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\SoftDeletingScope;
 
 class KategoriResource extends Resource
 {
+
     protected static ?string $model            = Kategori::class;
     protected static ?string $navigationIcon   = 'heroicon-o-collection';
     protected static ?string $navigationGroup  = 'Medisinsk';
     protected static ?string $modelLabel       = 'Kategori';
     protected static ?string $pluralModelLabel = 'Kategorier';
-    protected static ?int $navigationSort      = 2;
+    protected static ?int    $navigationSort   = 2;
 
-    protected static function getNavigationBadge(): ?string
+    protected static function getNavigationBadge() : ?string
     {
+
         return static::getModel()::count();
     }
 
-    public static function form(Form $form): Form
+    public static function form(Form $form) : Form
     {
+
         return $form
             ->schema([
-                Forms\Components\TextInput::make('kategori')
-                    ->required()
-                    ->maxLength(191),
+                TextInput::make('kategori')
+                         ->required()
+                         ->maxLength(191),
             ]);
     }
 
-    public static function table(Table $table): Table
+    /**
+     * @throws \Exception
+     */
+    public static function table(Table $table) : Table
     {
+
         return $table
             ->columns([
-                Tables\Columns\TextColumn::make('kategori'),
-                Tables\Columns\TextColumn::make('deleted_at')
-                    ->dateTime(),
-                Tables\Columns\TextColumn::make('created_at')
-                    ->dateTime(),
-                Tables\Columns\TextColumn::make('updated_at')
-                    ->dateTime(),
+                TextColumn::make('kategori'),
+                TextColumn::make('deleted_at')
+                          ->dateTime(),
+                TextColumn::make('created_at')
+                          ->dateTime(),
+                TextColumn::make('updated_at')
+                          ->dateTime(),
             ])
             ->filters([
                 Tables\Filters\TrashedFilter::make(),
@@ -62,29 +71,32 @@ class KategoriResource extends Resource
                 Tables\Actions\RestoreBulkAction::make(),
             ]);
     }
-    
-    public static function getRelations(): array
+
+    public static function getRelations() : array
     {
+
         return [
             //
         ];
     }
-    
-    public static function getPages(): array
+
+    public static function getPages() : array
     {
+
         return [
             'index'  => Pages\ListKategoris::route('/'),
             'create' => Pages\CreateKategori::route('/create'),
             'view'   => Pages\ViewKategori::route('/{record}'),
             'edit'   => Pages\EditKategori::route('/{record}/edit'),
         ];
-    }    
-    
-    public static function getEloquentQuery(): Builder
+    }
+
+    public static function getEloquentQuery() : Builder
     {
+
         return parent::getEloquentQuery()
-            ->withoutGlobalScopes([
-                SoftDeletingScope::class,
-            ]);
+                     ->withoutGlobalScopes([
+                         SoftDeletingScope::class,
+                     ]);
     }
 }
