@@ -21,7 +21,7 @@ class Ansatte extends BaseWidget
 
     public function getTableRecordKey(Model $record): string
     {
-        return uniqid();
+        return $record;
     }
 
     protected function getTableQuery(): Builder
@@ -48,7 +48,7 @@ class Ansatte extends BaseWidget
             Tables\Columns\TextColumn::make('email')
                 ->label('E-post')
                 ->limit(10)
-                ->tooltip(fn (Model $record): string => "{$record->email}"),
+                ->tooltip(fn (Model $record): string => "$record->email"),
             Tables\Columns\TextColumn::make('phone')
                 ->label('Telefon')
                 ->default('12345678'),
@@ -67,9 +67,7 @@ class Ansatte extends BaseWidget
                         )
                         ->where('unavailable', '!=', 1)->sum('totalt');
 
-                    $hours = sprintf('%02d', intdiv($minutes, 60)) . ':' . (sprintf('%02d', $minutes % 60));
-
-                    return $hours;
+                    return sprintf('%02d', intdiv($minutes, 60)).':'.(sprintf('%02d', $minutes % 60));
                 })
                 ->default('0'),
         ];
