@@ -33,9 +33,9 @@ class BestillUtstyr extends Mailable
     /**
      * Get the message envelope.
      *
-     * @return \Illuminate\Mail\Mailables\Envelope
+     * @return Envelope
      */
-    public function envelope()
+    public function envelope() : Envelope
     {
         return new Envelope(
             from: new Address('tor@trivera.net', 'Tor J. Rivera'),
@@ -49,9 +49,9 @@ class BestillUtstyr extends Mailable
     /**
      * Get the message content definition.
      *
-     * @return \Illuminate\Mail\Mailables\Content
+     * @return Content
      */
-    public function content()
+    public function content() : Content
     {
         return new Content(
             view: 'emails.bestill',
@@ -63,7 +63,7 @@ class BestillUtstyr extends Mailable
      *
      * @return array
      */
-    public function attachments()
+    public function attachments() : array
     {
         return [];
     }
