@@ -1,12 +1,16 @@
 <?php
-
 namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Notifications\Notifiable;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
+
+/**
+ * @mixin IdeHelperTestResults
+ */
 class TestResults extends Model
 {
     use HasFactory;
@@ -24,7 +28,7 @@ class TestResults extends Model
         'dato' => 'datetime',
     ];
 
-    public function tests()
+    public function tests(): BelongsTo
     {
         return $this->belongsTo(Tests::class, 'testsID');
     }
