@@ -3,10 +3,14 @@
 namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Notifications\Notifiable;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 
+/**
+ * @mixin IdeHelperTests
+ */
 class Tests extends Model
 {
     use HasFactory;
@@ -26,8 +30,8 @@ class Tests extends Model
         'ovelser' => 'array',
     ];
 
-    public function testResults()
+    public function testResults(): HasMany
     {
-        return $this->hasMany(TestResult::class);
+        return $this->hasMany(TestResults::class);
     }
 }
