@@ -5,6 +5,7 @@ namespace App\Models;
 use Cache;
 use Carbon\Carbon;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Database\Eloquent\Builder;
 use App\Notifications\UserUnavailable;
@@ -14,6 +15,7 @@ use Illuminate\Notifications\DatabaseNotification;
 
 /**
  * @method thisYear()
+ * @mixin IdeHelperKategori
  */
 class Kategori extends Model
 {
@@ -23,7 +25,7 @@ class Kategori extends Model
     protected $table = 'kategori';
 
     public $timestamps = true;
-    // protected $dates      = ['fra_dato', 'til_dato'];
+
     protected $fillable   = [
         'kategori',
     ];
@@ -32,14 +34,11 @@ class Kategori extends Model
     {
 
         parent::boot();
-        // static::addGlobalScope('unavailable', function (Builder $builder) {
 
-        //     $builder->where('unavailable', '=', 0);
-        // });
     }
 
     /**
-     * @return \Illuminate\Database\Eloquent\Relations\HasMany
+     * @return HasMany
      */
     public function utstyr()
     {
